fixed chris's errors. ;)

+++ trunk/source/kernel/idt.c
@@ -36,7 +36,7 @@ void idt_set_gate(BYTE num, DWORD base, WORD sel, BYTE flags)
 void idt_install(void)
 {
 	idtp.limit=((sizeof(struct idt_entry)*256)-1);
-	idtp.base=&idt;
+	idtp.base=(UINT)&idt;
 	_memset(&idt,0,(sizeof(struct idt_entry)*256));
 	idt_load();
 }
